% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dls.R
\name{doubleLassoSelect}
\alias{doubleLassoSelect}
\title{A function implements the Double Lasso Selection}
\usage{
doubleLassoSelect(df, outcome, treatment, test, k = 15)
}
\arguments{
\item{df}{Accepts \code{data.frame} and \code{data.table}. The data frame must contain all the variables specified in \code{outcome}, \code{treatment}, and \code{test}.}

\item{outcome}{Accepts single \code{character} value. It cannot be an empty character. The character specifies the outcome variable's name, which will be searched in the column names of provided data frame.}

\item{treatment}{Accepts single \code{character} value or a \code{character vector}. It specifies the treatment variable's name(s), which will be searched in the column names of provided data frame.
The treatment variables are those variables will NOT go through the selection and will be included in the final output data set.
This parameter accepts empty \code{character}, which implies no treatment variable to be included in the process.}

\item{test}{Accepts single empty \code{character} or a \code{character vector} with a length >= 2 (restricted by the \code{glmet} package). It specifies the test variable's name(s), which will be searched in the column names of provided data frame.
The test variables are those covariates will go through the selection and may or may not be included in the final data set.
This parameter accepts empty \code{character}, which implies performing selection on all variables except for the outcome and treatment variables.}

\item{k}{Accepts a \code{numeric} value. This is the number of times \code{lambda} being updated. The \code{lambda} here is a parameter used in lasso regression to represent the degree of regularization.
You do not have to adjust this value in most situations. The default value is suggested by the paper specified in the package reference.}
}
\value{
This function returns a data frame (\code{data.table}) with selected variables.
}
\description{
This function implements Double Lasso Selection on a specified data frame, with specified treatment variables to be included in the final model and covariates to be tested via the selection process.
}
\examples{
#Fetch data for demonstration
data(mtcars)

#Input example 1:
#Character vectors as `treatment` and `test` input with an interaction term
outcome <- "mpg"
treatment <- c("cyl", "hp")
test <- c("drat", "disp", "vs", "cyl:hp")

#Input example 2:
#Empty character as `treatment` and `test` input
outcome <- "mpg"
treatment <- ""
test <- ""

#Acquire the selected data frame
DT_select <- doubleLassoSelect(df=mtcars, outcome=outcome, treatment=treatment, test=test)

#Implement a linear model after the selection
model_lm <- lm(as.formula(sprintf("`\%s` ~ .", outcome)), data=DT_select)
summary(model_lm)

}
\keyword{double}
\keyword{lasso}
\keyword{selection}
\keyword{variable}
