% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlsr.R
\docType{package}
\name{dlsr-package}
\alias{dlsr}
\alias{dlsr-package}
\title{dlsr: What the Package Does (one line, title case)}
\description{
What the package does (one paragraph).
}
\details{
The only function you're likely to need from roxygen2 is [roxygenize()].
Otherwise refer to the vignettes to see how to format the documentation.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
